Refactor inf.py into modular components
